feat: connect api
API from edamam has been connected to the application. It should be
noted that api_key & app_id are not passed in headers, but in params and
have the name app_key, app_id, respectively. It sounds completely wrong,
but purely according to my observations, this api does not support
api-key in headers.